Move PubkeyMatch rule test to separate file
